requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
nltk==3.8.1
scikit-learn==1.3.2
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

